+++ src/main/java/CoinMerge/coinMergeSpring/news/dto/newsDTO.java
@@ -1,9 +1,25 @@
-package CoinMerge.coinMergeSpring.news;
+package CoinMerge.coinMergeSpring.news.dto;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Table(name = "news")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class newsDTO {
+    @Id @Column(name = "title")
     String title;
+    @Column(name = "content")
     String content;
+    @Column(name = "link")
     String link;
+    @Column(name = "date")
     String date;
 
     public newsDTO(String title, String content, String link, String date) {
