+++ src/main/java/CoinMerge/coinMergeSpring/news/newsRepository.java
@@ -1,0 +1,7 @@
+package CoinMerge.coinMergeSpring.news;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+@Repository
+public class newsRepository  {
+}
